Fix namespaces for types moved to ConnectedMode
Part of #4312
