+++ src/ConnectedMode.UnitTests/Binding/FixedStepsProgressTests.cs
@@ -19,11 +19,9 @@
  */
 
 using System;
-using FluentAssertions;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using SonarLint.VisualStudio.Integration.Binding;
+using SonarLint.VisualStudio.ConnectedMode.Binding;
 
-namespace SonarLint.VisualStudio.Integration.UnitTests.Binding
+namespace SonarLint.VisualStudio.ConnectedMode.UnitTests.Binding
 {
     [TestClass]
     public class FixedStepsProgressTests
