+++ src/ConnectedMode/Binding/FixedStepsProgress.cs
@@ -20,7 +20,7 @@
 
 using System;
 
-namespace SonarLint.VisualStudio.Integration.Binding
+namespace SonarLint.VisualStudio.ConnectedMode.Binding
 {
     /// <summary>
     /// Data class used for reporting progress through a fixed number of steps
