+++ src/Integration.UnitTests/Persistence/SolutionBindingCredentialsLoaderTests.cs
@@ -23,6 +23,7 @@ using FluentAssertions;
 using Microsoft.Alm.Authentication;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
+using SonarLint.VisualStudio.ConnectedMode.Persistence;
 using SonarLint.VisualStudio.Core.Binding;
 using SonarLint.VisualStudio.Integration.Persistence;
 using SonarQube.Client.Helpers;
