+++ src/Integration/Binding/BindingWorkflow.cs
@@ -21,6 +21,7 @@
 using System;
 using System.Diagnostics;
 using System.Threading;
+using SonarLint.VisualStudio.ConnectedMode.Binding;
 using SonarLint.VisualStudio.Integration.Progress;
 using SonarLint.VisualStudio.Integration.Resources;
 using SonarLint.VisualStudio.Progress.Controller;
