+++ src/Integration/Binding/IBindingWorkflowExecutor.cs
@@ -18,6 +18,8 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
+using SonarLint.VisualStudio.ConnectedMode.Binding;
+
 namespace SonarLint.VisualStudio.Integration.Binding
 {
     // Test only interface
