+++ src/Integration/NewConnectedMode/IConfigurationPersister.cs
@@ -21,6 +21,7 @@
 using System;
 using System.ComponentModel.Composition;
 using System.IO;
+using SonarLint.VisualStudio.ConnectedMode.Binding;
 using SonarLint.VisualStudio.Core.Binding;
 using SonarLint.VisualStudio.Integration.Persistence;
 using SonarLint.VisualStudio.Integration.UnintrusiveBinding;
