+++ src/TestInfrastructure/Framework/ConfigurableSectionController.cs
@@ -19,7 +19,7 @@
  */
 
 using System.Windows.Input;
-using SonarLint.VisualStudio.Integration.Binding;
+using SonarLint.VisualStudio.ConnectedMode.Binding;
 using SonarLint.VisualStudio.Integration.Progress;
 using SonarLint.VisualStudio.Integration.TeamExplorer;
 using SonarLint.VisualStudio.Integration.WPF;
